% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_K.R
\name{calc_allocation_prob}
\alias{calc_allocation_prob}
\title{Calculate the allocation probability}
\usage{
calc_allocation_prob(y, w, mu, kappa)
}
\arguments{
\item{y}{data vector}

\item{w}{a scalar weight}

\item{mu}{a scalar class mean}

\item{kappa}{a scalar inverse variance}
}
\description{
Calculate the allocation probability
}

