% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mu.R
\name{calc_C}
\alias{calc_C}
\title{Calculate the C matrix for DPP}
\usage{
calc_C(mu, theta, tau)
}
\arguments{
\item{mu}{vector of means}

\item{theta}{hyperparameter, a scalar}

\item{tau}{hyperparameter, a scalar}
}
\description{
Calculate the C matrix for DPP
}

