% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rho.R
\name{calc_detJ}
\alias{calc_detJ}
\title{Calculate determinant of Jacobian}
\usage{
calc_detJ(w_big, w_small, kappa_big, kappa_small, r)
}
\arguments{
\item{w_big}{weight vector from model with greater $K$}

\item{w_small}{weight vector from model with smaller $K$}

\item{kappa_big}{kappa vector from model with greater $K$}

\item{kappa_small}{kappa vector from model with smaller $K$}
}
\description{
Calculate determinant of Jacobian
}

