% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rho.R
\name{calc_kappa_ratio}
\alias{calc_kappa_ratio}
\title{Calculate kappa ratio}
\usage{
calc_kappa_ratio(kappa_big, kappa_small, a, b)
}
\arguments{
\item{kappa_big}{kappa vector from the model with a greater $K$}

\item{kappa_small}{kappa vector from the model with smaller $K$}

\item{a}{hyperparameter a}

\item{b}{hyperparameter b}
}
\description{
Calculate kappa ratio
}

