% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rho.R
\name{calc_lik_ratio}
\alias{calc_lik_ratio}
\title{Calculate likelihood ratio}
\usage{
calc_lik_ratio(s_big = s_big, s_small = s_small, w_big = w_big,
  w_small = w_small, mu_big = mu_big, mu_small = mu_small,
  kappa_big = kappa_big, kappa_small = kappa_small, y = y)
}
\arguments{
\item{s_big}{s for big model}

\item{s_small}{s for small model}

\item{w_big}{w for big model}

\item{w_small}{w for small model}

\item{mu_big}{mu for big model}

\item{mu_small}{mu for small model}

\item{kappa_big}{kappa for big model}

\item{kappa_small}{kappa for small model}

\item{y}{data vector}
}
\description{
Calculate likelihood ratio
}

