% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mu.R
\name{calc_little_c}
\alias{calc_little_c}
\title{little c, when we're using squared exponential kernel}
\usage{
calc_little_c(mu1, mu2, theta)
}
\description{
little c, when we're using squared exponential kernel
}

