% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rho.R
\name{calc_mu_ratio}
\alias{calc_mu_ratio}
\title{Calculate mu ratio}
\usage{
calc_mu_ratio(mu_big, mu_small, theta = theta, tau = tau)
}
\arguments{
\item{mu_big}{mean vector for model with greater $K$}

\item{mu_small}{mean vector for model with smaller $K$}

\item{theta}{hyperparameter theta for DPP}

\item{tau}{hyperparameter tau for DPP}
}
\description{
Calculate mu ratio
}

