% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rho.R
\name{calc_rho}
\alias{calc_rho}
\title{Calculate rho(omega_current, omega_proposal)}
\usage{
calc_rho(omega_small, omega_big, y, theta = theta, tau = tau)
}
\arguments{
\item{omega_small}{parameter vector from model with smaller $K$}

\item{omega_big}{parameter vector from model with bigger $K$}

\item{y}{a data vector}

\item{theta}{DPP hyperparameter}

\item{tau}{DPP hyperparameter}
}
\description{
Calculate rho(omega_current, omega_proposal)
}

