% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rho.R
\name{calc_w_ratio}
\alias{calc_w_ratio}
\title{Calculate w ratio}
\usage{
calc_w_ratio(w_big, w_small, s_big, s_small, delta = 1)
}
\arguments{
\item{w_big}{w vector of weights from model with greater $K$}

\item{w_small}{w vector of weights from model with smaller $K$}
}
\description{
Calculate w ratio
}

