% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_K.R
\name{define_big_s}
\alias{define_big_s}
\title{Define a s vector for the model with a longer mu vector}
\usage{
define_big_s(s, ind1, ind2, y, w_new, mu_new, kappa_new)
}
\description{
Define a s vector for the model with a longer mu vector
}

