% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_K.R
\name{define_big_w}
\alias{define_big_w}
\title{Define a w weight vector that is one longer than the inputted weight vector}
\usage{
define_big_w(w, alpha, ind1, ind2 = length(w) + 1)
}
\description{
Define a w weight vector that is one longer than the inputted weight vector
}

