% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_K.R
\name{define_extra_parameters}
\alias{define_extra_parameters}
\title{Define three parameters, alpha, beta and r for dimension-matching purposes in RJ MCMC}
\usage{
define_extra_parameters()
}
\description{
Define three parameters, alpha, beta and r for dimension-matching purposes in RJ MCMC
}

