% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_K.R
\name{define_small_s}
\alias{define_small_s}
\title{Define a s vector for the smaller model, ie the model with shorter mu vector}
\usage{
define_small_s(s, mu, ind1, ind2)
}
\description{
Define a s vector for the smaller model, ie the model with shorter mu vector
}

