% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_parameter_vector.R
\name{order_parameter_vector}
\alias{order_parameter_vector}
\title{Order parameter vector (mu, sigma, w) according to entries in mu vector}
\usage{
order_parameter_vector(mu, sigma, w, s)
}
\arguments{
\item{mu}{a mean vector}

\item{sigma}{a standard deviation vector}

\item{w}{a weight vector}

\item{s}{an allocation vector}
}
\description{
Order parameter vector (mu, sigma, w) according to entries in mu vector
}

