% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mu.R
\name{update_mu}
\alias{update_mu}
\title{Update $\mu$ in Gibbs sampling}
\usage{
update_mu(y, s, mu, sigma, tau, theta)
}
\arguments{
\item{y}{data vector}

\item{s}{vector of class assignments}

\item{mu}{vector of class means}

\item{sigma}{vector of class standard deviations}

\item{tau}{hyperparameter}
}
\description{
Update $\mu$ in Gibbs sampling
}

