% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sigma.R
\name{update_sigma}
\alias{update_sigma}
\title{Update sigma in Gibbs sampling}
\usage{
update_sigma(y, mu, s, a, b)
}
\arguments{
\item{y}{data vector}

\item{mu}{length-K vector of cluster means}

\item{s}{vector of cluster assignments, same length as y}

\item{a}{hyperparameter, a scalar}

\item{b}{hyperparameter, a scalar}
}
\value{
updated sigma, a vector of cluster standard deviations
}
\description{
Update sigma in Gibbs sampling
}
\examples{
update_sigma(rnorm(6), mu = c(0, 1, 1), s = c(1,1,2,2,3,3), a = 1, b = 1)
}

