% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tau_gamma.R
\name{update_tau_gamma}
\alias{update_tau_gamma}
\title{Update tau in Gibbs sampling}
\usage{
update_tau_gamma(tau, theta, mu, shape = 3, scale = 20, exponent = 0:199)
}
\description{
Update tau in Gibbs sampling
}

