% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_theta_gamma.R
\name{update_theta_gamma}
\alias{update_theta_gamma}
\title{Update $theta$ in Gibbs sampling}
\usage{
update_theta_gamma(theta, tau, mu, shape = 3, scale = 20,
  exponent = 0:199)
}
\arguments{
\item{theta}{scalar}

\item{tau}{scalar}

\item{mu}{vector of means}

\item{shape}{hyperparameter, prior shape for theta}

\item{scale}{hyperparameter, prior scale for theta}
}
\description{
Update $theta$ in Gibbs sampling
}

