% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_w.R
\name{update_w}
\alias{update_w}
\title{Update w in Gibbs sampling}
\usage{
update_w(mu, s, delta = 1)
}
\arguments{
\item{s}{a vector with one entry per observed data point of cluster assignments}

\item{delta}{a scalar hyperparameter for dirichlet}
}
\value{
w a weight vector of length $K$
}
\description{
Update w in Gibbs sampling
}
\examples{
update_w(1:4, c(1,1,2,2,3,3))

}

