% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_K.R
\name{calc_rho}
\alias{calc_rho}
\title{Calculate rho, the acceptance probability for the update K step}
\usage{
calc_rho(y, omega_small, omega_big, ind1, ind2, a, b, alpha, beta, r, delta)
}
\arguments{
\item{y}{data vector}

\item{omega_small}{list containing the current state of parameter vector}

\item{omega_big}{list containing the proposed state of parameter vector}

\item{ind1}{indicator denoting which cluster is affected}

\item{ind2}{indicator denoting other affected cluster}

\item{a}{hyperparameter scalar}

\item{b}{a scalar hyperparameter}

\item{alpha}{scalar parameter for dimension-matching}

\item{beta}{scalar parameter for dimension-matching}

\item{r}{scalar parameter for dimension-matching}

\item{delta}{scalar hyperparameter}
}
\description{
Calculate rho, the acceptance probability for the update K step
}

