% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_K.R
\name{calc_rho}
\alias{calc_rho}
\title{Calculate rho, the acceptance probability for the update K step}
\usage{
calc_rho(y, omega_small, omega_big, ind, a, b, alpha, beta, r, delta)
}
\arguments{
\item{omega_small}{list containing the current state of parameter vector}

\item{omega_big}{list containing the proposed state of parameter vector}
}
\description{
Calculate rho, the acceptance probability for the update K step
}

