% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_theta.R
\name{update_theta}
\alias{update_theta}
\title{Update $theta$ in Gibbs sampling}
\usage{
update_theta(theta, tau, mu, a1 = 0, b1 = 50, exponent = 0:199)
}
\arguments{
\item{theta}{scalar}

\item{tau}{scalar}

\item{mu}{vector of means}

\item{a1}{hyperparameter, prior mean for theta}

\item{b1}{hyperparameter, prior sd for theta}
}
\description{
Update $theta$ in Gibbs sampling
}

