% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_w.R
\name{update_w}
\alias{update_w}
\title{Update w in Gibbs sampling}
\usage{
update_w(delta, s)
}
\arguments{
\item{delta}{a K-length vector of hyperparameters}

\item{s}{a vector with one entry per observed data point of cluster assignments}
}
\value{
w a weight vector of length $K$
}
\description{
Update w in Gibbs sampling
}
\examples{
update_w(c(1,1,1), c(1,1,2,2,3,3))

}

